+++ backend_v3/src/services/transferServices.ts
@@ -1,10 +1,10 @@
+import * as transfer from '../database/transfer'
 import { Transfer } from '../types'
-import * as transfer from '../controllers/transfer.controller'
 import { log } from '../utils/functions'
 
 export const getTransfersFromVehicle = async (vehicleId: string, debug = false): Promise<Transfer[] | null> => {
   try {
-    const transfers: Transfer[] | null = await transfer.findByIdVehicle(vehicleId, debug)
+    const transfers: Transfer[] | null = await transfer.findByVehicleId(vehicleId, debug)
     return transfers
   } catch (err: any) {
     debug && log('ERROR', err.message)
